+++ flooding_worker/management/commands/perform_task.py
@@ -4,9 +4,9 @@ import logging #, threading, time, datetime, random, math
 
 from django.core.management.base import BaseCommand
 from flooding import settings
-from lizard_flooding_worker.perform_task import perform_task
+from flooding_worker.perform_task import perform_task
 
-log = logging.getLogger('nens.lizard.flooding.worker.management.command.perform_task')
+log = logging.getLogger('flooding.worker.management.command.perform_task')
 
 
 class Command(BaseCommand):