+++ flooding_worker/management/commands/start_scenario_new.py
@@ -4,12 +4,12 @@
 from optparse import make_option
 
 from django.core.management.base import BaseCommand
-from lizard_flooding_worker.worker.action_workflow import ActionWorkflow
-from lizard_flooding_worker.worker.broker_connection import BrokerConnection
-from lizard_flooding_worker.worker.message_logging_handler import AMQPMessageHandler
+from flooding_worker.worker.action_workflow import ActionWorkflow
+from flooding_worker.worker.broker_connection import BrokerConnection
+from flooding_worker.worker.message_logging_handler import AMQPMessageHandler
 
 import logging
-log = logging.getLogger("lizard-flooding.management.start_scenario")
+log = logging.getLogger("flooding.management.start_scenario")
 
 class Command(BaseCommand):