+++ flooding_worker/management/commands/task_worker_new.py
@@ -4,13 +4,13 @@
 from optparse import make_option
 
 from django.core.management.base import BaseCommand
-from lizard_flooding_worker.worker.worker import Worker
-from lizard_flooding_worker.worker.action_task import ActionTask
-from lizard_flooding_worker.worker.broker_connection import BrokerConnection
-from lizard_flooding_worker.worker.message_logging_handler import AMQPMessageHandler
+from flooding_worker.worker.worker import Worker
+from flooding_worker.worker.action_task import ActionTask
+from flooding_worker.worker.broker_connection import BrokerConnection
+from flooding_worker.worker.message_logging_handler import AMQPMessageHandler
 
 import logging
-log = logging.getLogger("lizard-flooding.management.logging_worker")
+log = logging.getLogger("flooding.management.logging_worker")
 
 
 class Command(BaseCommand):